requirements.txt
requests~=2.25.1
python-dotenv~=0.18.0
twilio~=6.61.0
Flask~=1.1.4
fastapi~=0.68.1